feat: add test for /tweets/
